- much work on making file->open work.
- right now I can open a air_coil file (its hard coded to only do air_coil).

TODO:

finish up the wcalc_loadsave stuff to determine file type.

finish the wcalc.c stuff that goes with that.

unify the code that puts the filename in the title bar.


-----
this code isn't done, but I won't get a chance to work on it for some time
so commit this to avoid losing it.