[gtk] Fix up some print/pdf export issues
Not quite sure how this worked before, but it wasn't working with
the parallel wires GUI so give it a bit of an overhaul.  Noted that
all initialization of the wcalc superclass was being done in the
individual subclass initializations.  Update this to put the
common init stuff into a common location.  This is still OOP in
C but it is better than it was.

Remove some homegrown code dating back a few decades for messing with
filename strings.  Now keep track of the full path filename, the
directory portion, the file portion, and the file basename portion.
This allows easier/better support for the PDF file name for PDF
export.