[gtk] Huge cleanup in prep for clean 2.24 support
This is a huge commit removing most, but not all yet, usage
of deprecated gtk code.  This is a step towards prep for
porting to gtk3.

Use this to compile with more strict checks in place.
make CFLAGS+="-DGTK_DISABLE_SINGLE_INCLUDES -DGDK_DISABLE_DEPRECATED -DGTK_DISABLE_DEPRECATED -DGSEAL_ENABLE"