[libwcalc] Add some additional comments in the coplanar code
We have a few open issues in the coplanar model.  This adds some
comments but no actual resolution.

- the model with ground plane doesn't converge to microstrip with
very large spacing.  Note that the model takes an extremely large gap
(much much larger than what makes sense) to converge to the infinite
gap case.  I expect that we need to put in some as yet to be determined
limit checks for model validity range.

- add a comparison calculation from a text although the text is a little
unclear about it, the equations appear to be for the *no* ground plane
case.