[win32] Add wcalc_icon.ico
Add the win32 icon file.  We don't like rebuilding this as it needs
tgif that isn't as commonly installed and on cygwin in particular
it is more painful due to the need for a working X server.  Plus this
is a low traffic file so go ahead and put into git.