[libwcalc] Work on improving the grounded CPW analysis
For CPW with a bottom side ground, calculate the impedance and
effective dielectric constant for a microstrip of the same dimensions.
Use this as a limiting value for large gaps.  In the vicinity of
where CPW and microstrip values approach each other, use a spline
interpolation for a smoother transition.

While this is an improvement in the large gap case, we still need
further improvements to account for metal thickness in the case
with bottom side ground.