[ci] Overhaul the ci pipeline and get it all working
This has lots of fixes including a more unified flow where differences
between platforms is controlled with conditional steps instead of
maintaining 2 unique pipelines.

On the macOS side, I wasn't able to find xvfb-run so downloaded
a copy of the upstream sources from x.org and apply some minor
tweaks (`gfmt` instead of `fmt`, more verbose output).

There was also an issue when running `make distcheck` on macOS.
The desktop icons, currently drawn with tgif, get exported to xpm
and on macOS, two of the color names then caused errors when fed
to the ImageMagick `convert` utility.  This appears to be related
to which `rgb.txt` file is found (or not found).  Explicitly setting
RGBDEF to point at the /opt/X11/... version seems to have taken
care of this.